myproject/requirements.txt
Django==2.1.3
PyMySQL==0.9.2
djangorestframework==3.9.1
python-dateutil==2.7.5
python-mimeparse==1.6.0
pytz==2018.9
six==1.12.0
django-widget-tweaks==1.4.3
